Move routines from brcm_patchram_plus_usb.c to brcm_usb.c
